Refactor: 패키지 수정, filter 단순화
패키지에 jwt 삭제, filter 구조 단순화로 security에서 잡도록 수정, config 이동
