+++ src/main/java/nbc/chillguys/nebulazone/infra/security/exception/JwtTokenErrorCode.java
@@ -1,4 +1,4 @@
-package nbc.chillguys.nebulazone.infra.security.jwt.exception;
+package nbc.chillguys.nebulazone.infra.security.exception;
 
 import org.springframework.http.HttpStatus;
