+++ src/main/java/nbc/chillguys/nebulazone/infra/security/filter/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package nbc.chillguys.nebulazone.infra.security.jwt.filter;
+package nbc.chillguys.nebulazone.infra.security.filter;
 
 import java.io.IOException;
 
@@ -14,10 +14,10 @@ import jakarta.servlet.http.HttpServletResponse;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import nbc.chillguys.nebulazone.domain.auth.vo.AuthUser;
-import nbc.chillguys.nebulazone.infra.security.jwt.JwtUtil;
-import nbc.chillguys.nebulazone.infra.security.jwt.constant.JwtConstants;
-import nbc.chillguys.nebulazone.infra.security.jwt.filter.exception.JwtFilterErrorCode;
-import nbc.chillguys.nebulazone.infra.security.jwt.filter.exception.JwtFilterException;
+import nbc.chillguys.nebulazone.infra.security.JwtUtil;
+import nbc.chillguys.nebulazone.infra.security.constant.JwtConstants;
+import nbc.chillguys.nebulazone.infra.security.filter.exception.JwtFilterErrorCode;
+import nbc.chillguys.nebulazone.infra.security.filter.exception.JwtFilterException;
 
 @RequiredArgsConstructor
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
@@ -27,40 +27,20 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 	@Override
 	protected void doFilterInternal(@NonNull HttpServletRequest request, @NonNull HttpServletResponse response,
 		@NonNull FilterChain filterChain) throws ServletException, IOException {
-		String uri = request.getRequestURI();
-
-		if (JwtConstants.PUBLIC_URLS.stream().anyMatch(uri::startsWith)) {
-			filterChain.doFilter(request, response);
-			return;
-		}
-
 		try {
 			String accessHeader = request.getHeader(JwtConstants.AUTH_HEADER);
-			if (accessHeader == null || accessHeader.isBlank()) {
-				throw new JwtFilterException(JwtFilterErrorCode.EMPTY_TOKEN);
-			}
-
-			String accessToken = resolveToken(accessHeader);
-			if (jwtUtil.isTokenExpired(accessToken)) {
-				String refreshHeader = request.getHeader(JwtConstants.REFRESH_HEADER);
-				if (refreshHeader == null || refreshHeader.isBlank()) {
+			if (accessHeader != null && !accessHeader.isBlank()) {
+				String accessToken = resolveToken(accessHeader);
+				if (jwtUtil.isTokenExpired(accessToken)) {
 					throw new JwtFilterException(JwtFilterErrorCode.TOKEN_EXPIRED);
 				}
 
-				String refreshToken = resolveToken(refreshHeader);
-				if (jwtUtil.isTokenExpired(refreshToken)) {
-					throw new JwtFilterException(JwtFilterErrorCode.TOKEN_EXPIRED);
-				}
+				AuthUser authUser = jwtUtil.getAuthUserFromToken(accessToken);
+				Authentication authentication = new UsernamePasswordAuthenticationToken(authUser, accessToken,
+					authUser.getAuthorities());
 
-				accessToken = jwtUtil.regenerateAccessToken(refreshToken);
-				response.setHeader(JwtConstants.AUTH_HEADER, accessToken);
+				SecurityContextHolder.getContext().setAuthentication(authentication);
 			}
-
-			AuthUser authUser = jwtUtil.getAuthUserFromToken(accessToken);
-			Authentication authentication = new UsernamePasswordAuthenticationToken(authUser, accessToken,
-				authUser.getAuthorities());
-
-			SecurityContextHolder.getContext().setAuthentication(authentication);
 
 			filterChain.doFilter(request, response);
 		} catch (JwtFilterException e) {
